--- Example/chains/exactBB_lmax500.covmat
@@ -1,3 +1,0 @@
-# nt r
-  2.5431283E-03   7.1978162E-06
-  7.1978162E-06   2.8729585E-07
