--- Example/chains/exactBB_lmax500.input.yaml
@@ -1,60 +1,0 @@
-theory:
-  camb:
-    extra_args:
-      bbn_predictor: PArthENoPE_880.2_standard.dat
-      halofit_version: mead
-      lens_potential_accuracy: 1
-      NonLinear: NonLinear_both
-      max_l: 2700
-      WantTransfer: true
-      Transfer.high_precision: true
-      parameterization: 2
-      num_nu_massless: 2.046
-      share_delta_neff: true
-      YHe: 0.2454006
-      pivot_tensor: 0.01
-      num_massive_neutrinos: 1
-      theta_H0_range:
-      - 20
-      - 100
-likelihood:
-  BB: true
-params:
-  As: 2.100549e-09
-  ns: 0.9660499
-  ombh2: 0.0223828
-  omch2: 0.1201075
-  omnuh2: 0.0006451439
-  H0: 67.32117
-  tau: 0.05430842
-  nt:
-    latex: n_t
-    prior:
-      max: 5
-      min: -5
-    proposal: 0.1
-    ref:
-      dist: norm
-      loc: 0.1
-      scale: 0.001
-  r:
-    latex: r_{0.01}
-    prior:
-      max: 3
-      min: 1.0e-05
-    proposal: 0.0002
-    ref:
-      dist: norm
-      loc: 0.02
-      scale: 0.0005
-  r005:
-    derived: 'lambda r, nt, ns: r * (0.05/0.01)**(nt - ns + 1)'
-    min: 0
-    max: 3
-    latex: r_{0.05}
-sampler:
-  mcmc:
-    Rminus1_cl_stop: 0.2
-    Rminus1_stop: 0.01
-output: exactBB_lmax500
-stop-at-error: true
