--- Example/chains/exactBB_lmax500.progress
@@ -1,1 +1,0 @@
-#       N                  timestamp acceptance_rate         Rminus1      Rminus1_cl
