--- Example/chains/exactTTTEEEBB_lmax[1500, 1200, 500].checkpoint
@@ -1,6 +1,0 @@
-sampler:
-  mcmc:
-    converged: false
-    Rminus1_last: 6.709459126018196
-    burn_in: 0
-    mpi_size: 1
