--- Example/chains/exactTTTEEEBB_lmax[1500, 1200, 500].covmat
@@ -1,9 +1,0 @@
-# logA ns nt r ombh2 omch2 tau theta_MC_100
-2.862760661999632135e-06 2.153048582549491307e-07 -2.037084418853600377e-06 -1.652429867907647288e-07 1.333185440241335216e-08 -1.743701217652817290e-07 1.141480020934996837e-06 1.481721496943173433e-08
-2.153048582549491307e-07 5.740220081259920538e-06 -1.527103608599936538e-05 1.393336698229702095e-07 -5.047101508334674226e-08 -1.472240208490113930e-07 5.393398589674224003e-07 -7.842115983715230029e-08
--2.037084418853600377e-06 -1.527103608599936538e-05 3.516568066602370618e-03 1.301653678026341182e-05 -1.091602553424506738e-07 -2.954018534366108505e-06 5.662449380462547082e-06 -1.373829390086002933e-07
--1.652429867907647288e-07 1.393336698229702095e-07 1.301653678026341182e-05 2.610287343324101546e-07 -9.010401839841079739e-10 -5.797105527680898189e-08 3.215743383334468017e-08 -1.035750639609380730e-09
-1.333185440241335216e-08 -5.047101508334674226e-08 -1.091602553424506474e-07 -9.010401839841079739e-10 4.543659673159362150e-09 -1.144366125447496703e-09 -1.450531326638427869e-08 2.314431530679109732e-09
--1.743701217652817290e-07 -1.472240208490113930e-07 -2.954018534366109352e-06 -5.797105527680898189e-08 -1.144366125447496703e-09 8.957210136835323993e-08 -1.883672430753237712e-07 -2.148742834696872302e-09
-1.141480020934996837e-06 5.393398589674224003e-07 5.662449380462547082e-06 3.215743383334468017e-08 -1.450531326638427869e-08 -1.883672430753237712e-07 8.969276135675558511e-07 6.771841971341644708e-09
-1.481721496943173433e-08 -7.842115983715230029e-08 -1.373829390086003727e-07 -1.035750639609380730e-09 2.314431530679109732e-09 -2.148742834696872302e-09 6.771841971341644708e-09 3.116799343649793447e-08
