--- Example/chains/exactTTTEEEBB_lmax[1500, 1200, 500].input.yaml
@@ -1,128 +1,0 @@
-theory:
-  camb:
-    extra_args:
-      bbn_predictor: PArthENoPE_880.2_standard.dat
-      halofit_version: mead
-      lens_potential_accuracy: 1
-      NonLinear: NonLinear_both
-      max_l: 2700
-      WantTransfer: true
-      Transfer.high_precision: true
-      parameterization: 2
-      num_nu_massless: 2.046
-      share_delta_neff: true
-      YHe: 0.2454006
-      pivot_tensor: 0.01
-      num_massive_neutrinos: 1
-      theta_H0_range:
-      - 20
-      - 100
-likelihood:
-  TTTEEE: true
-  BB: true
-params:
-  As:
-    latex: A_\mathrm{s}
-    value: 'lambda logA: 1e-10*np.exp(logA)'
-  H0:
-    latex: H_0
-    max: 100
-    min: 20
-  cosmomc_theta:
-    derived: false
-    value: 'lambda theta_MC_100: 1.e-2*theta_MC_100'
-  logA:
-    drop: true
-    latex: \log(10^{10} A_\mathrm{s})
-    prior:
-      max: 3.91
-      min: 1.61
-    proposal: 0.001
-    ref:
-      dist: norm
-      loc: 3.044
-      scale: 0.0001
-  mnu: 0.06
-  ns:
-    latex: n_\mathrm{s}
-    prior:
-      max: 1.2
-      min: 0.8
-    proposal: 0.002
-    ref:
-      dist: norm
-      loc: 0.9660499
-      scale: 0.0004
-  nt:
-    latex: n_t
-    prior:
-      max: 5
-      min: -5
-    proposal: 0.1
-    ref:
-      dist: norm
-      loc: 0.1
-      scale: 0.001
-  r:
-    latex: r_{0.01}
-    prior:
-      max: 3
-      min: 1.0e-05
-    proposal: 0.0002
-    ref:
-      dist: norm
-      loc: 0.02
-      scale: 0.0005
-  r005:
-    derived: 'lambda r, nt, ns: r * (0.05/0.01)**(nt - ns + 1)'
-    min: 0
-    max: 3
-    latex: r_{0.05}
-  ombh2:
-    latex: \Omega_\mathrm{b} h^2
-    prior:
-      max: 0.1
-      min: 0.005
-    proposal: 0.0001
-    ref:
-      dist: norm
-      loc: 0.0223828
-      scale: 1.0e-05
-  omch2:
-    latex: \Omega_\mathrm{c} h^2
-    prior:
-      max: 0.99
-      min: 0.001
-    proposal: 0.0005
-    ref:
-      dist: norm
-      loc: 0.1201075
-      scale: 0.0001
-  tau:
-    latex: \tau_\mathrm{reio}
-    prior:
-      max: 0.8
-      min: 0.01
-    proposal: 0.003
-    ref:
-      dist: norm
-      loc: 0.05430842
-      scale: 0.0006
-  theta_MC_100:
-    drop: true
-    latex: 100\theta_\mathrm{MC}
-    prior:
-      max: 10
-      min: 0.5
-    proposal: 0.0002
-    ref:
-      dist: norm
-      loc: 1.04109
-      scale: 4.0e-05
-    renames: theta
-sampler:
-  mcmc:
-    Rminus1_cl_stop: 0.2
-    Rminus1_stop: 0.1
-output: exactTTTEEEBB_lmax[1500, 1200, 500]
-stop-at-error: true
