--- Example/chains/exactTTTEEEBB_lmax[1500, 1200, 500].progress
@@ -1,2 +1,0 @@
-#       N                  timestamp acceptance_rate         Rminus1      Rminus1_cl
- 256.0 2023-03-13T11:06:13.006550  0.205788  6.709459 NaN
