--- Example/chains/exactTTTEEE_lmax[1500, 1200].1.txt
@@ -1,12 +1,0 @@
-#        weight    minuslogpost            logA              ns           ombh2           omch2             tau    theta_MC_100              As              H0   minuslogprior minuslogprior__0            chi2    chi2__TTTEEE
-              2       1.7173604        3.044038      0.96582957     0.022368529      0.12005015     0.055035532       1.0411203   2.0989829e-09       67.226187     -0.43275148      -0.43275148       4.3002237       4.3002237
-              4       1.6508312        3.044038      0.96582957     0.022364635      0.12013417     0.054941716       1.0411313   2.0989829e-09       67.197327     -0.43275148      -0.43275148       4.1671654       4.1671654
-             14       1.1449568        3.044038      0.96582957     0.022221763      0.12054542     0.053671352       1.0414241   2.0989829e-09       67.034112     -0.43275148      -0.43275148       3.1554166       3.1554166
-              6       1.3852875       3.0430814      0.96732882     0.022221763      0.12054542     0.053671352       1.0414241    2.096976e-09       67.034112     -0.43275148      -0.43275148        3.636078        3.636078
-             13       1.7773565       3.0430814      0.96732882     0.022206193      0.12050597      0.05343826        1.041421    2.096976e-09       67.033975     -0.43275148      -0.43275148       4.4202159       4.4202159
-              1       2.2261628       3.0430814      0.96732882     0.022205307      0.12052217     0.053084292       1.0414159    2.096976e-09       67.025909     -0.43275148      -0.43275148       5.3178286       5.3178286
-             10       2.2950133       3.0427916      0.96250514     0.022205307      0.12052217     0.053084292       1.0414159   2.0963684e-09       67.025909     -0.43275148      -0.43275148       5.4555295       5.4555295
-              6         2.24093       3.0427916      0.96250514     0.022205485      0.12052831      0.05309815       1.0414151   2.0963684e-09       67.023676     -0.43275148      -0.43275148        5.347363        5.347363
-              4      0.98230885       3.0427916      0.96250514     0.022283303      0.12040857     0.053851114       1.0413199   2.0963684e-09       67.097667     -0.43275148      -0.43275148       2.8301207       2.8301207
-              9      0.91627751       3.0427916      0.96250514     0.022331226      0.12042842     0.053668312       1.0411992   2.0963684e-09       67.089997     -0.43275148      -0.43275148        2.698058        2.698058
-              3       1.0592676       3.0427916      0.96250514     0.022412581      0.12016689     0.052995934       1.0412593   2.0963684e-09       67.268304     -0.43275148      -0.43275148       2.9840382       2.9840382
