--- Example/chains/exactTTTEEE_lmax[1500, 1200].checkpoint
@@ -1,6 +1,0 @@
-sampler:
-  mcmc:
-    converged: false
-    Rminus1_last: .inf
-    burn_in: 0
-    mpi_size: 1
