--- Example/chains/exactTTTEEE_lmax[1500, 1200].progress
@@ -1,1 +1,0 @@
-#       N                  timestamp acceptance_rate         Rminus1      Rminus1_cl
