+++ lilit/__init__.py
@@ -1,0 +1,4 @@
+from .likelihood import LiLit
+
+__author__ = "Giacomo Galloni"
+__version__ = "1.0.0"
